refactor: component name and structure
