+++ src/components/Partnerships.tsx
