+++ src/global.scss
@@ -1031,7 +1031,7 @@ footer {
   }
 }
 
-// Testimonials
+// Partnerships
 .partners {
   text-align: center;
   min-height: 100vh;
